Move saving current card index to async task instead of loader.
Loaders are connected to an activity, so if saving index process isn’t
finished it would be done on next activity start, but current index on
activity start would be wrong. Async task has more chances to save
current index at a right time. It has no callbacks or connections to
activity so there would be no trouble to handle it.

Potentially the same thing could be applied for the deletion operations
(delete deck and delete card), I’m not really sure about this.