Fix bug with rotation and saving current card index.
After rotation and pressing “back” button some strange way index becomes
updated in database, but not updated in decks list. Of course, current
card index is a field in deck object and so deck list must be reloaded after
changes but the strange thing is in normal work with no-rotation
work-flow. This commit forces reloading in lists.